requirements.txt
numpy==1.26.4
opencv-python==4.8.1.78
Pillow==10.0.0
scikit-learn==1.3.2
scipy==1.11.3
matplotlib==3.8.0
joblib==1.3.2
imutils==0.5.4

#Jupyter notebook için:
notebook==7.0.4
ipython==8.15.0

pandas==2.1.1

python-dateutil==2.8.2
